+++ lib/data/repositories/invite_repository.dart
@@ -9,48 +9,15 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sqflite/sqflite.dart' as sqflite;
 
 abstract interface class InviteRepository {
-  /// Creates remotely an [Invite] with the provided [phoneNumber] and then saves it
-  /// to the local database;
-  ///
-  /// throws:
-  /// - [ApiPatientNotRegisteredException] when there is no patient registered
-  /// with the phone number that was passed;
-  /// - [ApiPatientAlreadyWithDoctorException] when the patient you are trying to
-  /// invite already is your patient;
-  /// - [ApiInviteAlreadySentException] when this invite was already sent to the patient;
-  /// - [ApiConnectionException] when the request suffers any connection problems;
-  /// - [ApiUnauthorizedException] when the response returns a status of 401 or 403;
-  ///
-  /// returns:
-  /// - the created [Invite];
-  Future<Invite> create(String phoneNumber);
+  Future<(Invite?, ApiException?)> create(String phoneNumber);
 
-  /// Fetches from the API the [Invite] list tied to this account and saves it
-  /// to the local database;
-  ///
-  /// throws:
-  /// - [ApiInvitesNotFoundException] when there are no invites tied to this account
-  /// - [ApiConnectionException] when the request suffers any connection problems;
-  /// - [ApiUnauthorizedException] when the response returns a status of 401 or 403;
-  Future<void> sync();
+  Future<ApiException?> sync();
 
-  /// returns:
-  /// - the [Invite] list;
   Future<List<Invite>> get();
 
-  /// Deletes the [Invite] with the provided [id];
-  ///
-  /// throws:
-  /// - [ApiConnectionException] when the request suffers any connection problems;
-  /// - [ApiUnauthorizedException] when the response returns a status of 401 or 403;
-  Future<void> delete(int id);
+  Future<ApiException?> delete(int id);
 
-  /// Accept the [Invite] with the provided [id], then deletes it locally;
-  ///
-  /// throws:
-  /// - [ApiConnectionException] when the request suffers any connection problems;
-  /// - [ApiUnauthorizedException] when the response returns a status of 401 or 403;
-  Future<void> accept(int id);
+  Future<ApiException?> accept(int id);
 
   /// Clears the table;
   Future<void> clear();
@@ -73,22 +40,27 @@ final class InviteRepositoryImpl implements InviteRepository {
   final sqflite.Database _db;
 
   @override
-  Future<Invite> create(String phoneNumber) async {
-    final createdInvite = await _api.createInvite(phoneNumber);
+  Future<(Invite?, ApiException?)> create(String phoneNumber) async {
+    final (createdInvite, err) = await _api.createInvite(phoneNumber);
+    if (err != null) return (null, err);
+
     await _db.insert(
       InviteEntity.tableName,
-      InviteMapper.toEntity(createdInvite).toMap(),
+      InviteMapper.toEntity(createdInvite!).toMap(),
       conflictAlgorithm: sqflite.ConflictAlgorithm.replace,
     );
-    return createdInvite;
+
+    return (createdInvite, null);
   }
 
   @override
-  Future<void> sync() async {
-    final invites = await _api.getInvites();
+  Future<ApiException?> sync() async {
+    final (invites, err) = await _api.getInvites();
+    if (err != null) return err;
+
     final batch = _db.batch();
     batch.delete(InviteEntity.tableName);
-    for (final invite in invites) {
+    for (final invite in invites!) {
       batch.insert(
         InviteEntity.tableName,
         InviteMapper.toEntity(invite).toMap(),
@@ -96,6 +68,8 @@ final class InviteRepositoryImpl implements InviteRepository {
       );
     }
     await batch.commit(noResult: true);
+
+    return null;
   }
 
   @override
@@ -117,25 +91,31 @@ final class InviteRepositoryImpl implements InviteRepository {
   }
 
   @override
-  Future<void> accept(int id) async {
-    await _api.acceptInvite(id);
+  Future<ApiException?> accept(int id) async {
+    final err = await _api.acceptInvite(id);
+    if (err != null) return err;
 
     await _db.delete(
       InviteEntity.tableName,
       where: "${InviteEntity.idColumn} = ?",
       whereArgs: [id],
     );
+
+    return null;
   }
 
   @override
-  Future<void> delete(int id) async {
-    await _api.deleteInvite(id);
+  Future<ApiException?> delete(int id) async {
+    final err = await _api.deleteInvite(id);
+    if (err != null) return err;
 
     await _db.delete(
       InviteEntity.tableName,
       where: "${InviteEntity.idColumn} = ?",
       whereArgs: [id],
     );
+
+    return null;
   }
 
   @override
